comment out thumbnail as it is not that useful at the moment
